+++ huoguoml/constants.py
@@ -17,5 +17,4 @@ HUOGUOML_DEFAULT_SERVICE_PORT: int = 5000
 HUOGUOML_DEFAULT_SERVICE_HOST: str = "127.0.0.1"
 
 # SERVER
-HUOGUOML_DEFAULT_SERVER_FOLDER: str = ".huoguoml"
 HUOGUOML_DEFAULT_SERVER_PORT: int = 8080
