+++ huoguoml/util/__init__.py
@@ -1,0 +1,3 @@
+"""
+The huoguoml.util module contains some utility functions used throughout the whole package
+"""
